Revert "late-night implementation of logout-if-session-is-invalid functionality"
This reverts commit c0ec5b4288c737c61c67e1a29da2dd9db396d58d.
